Añade componente para la cantidad de cada item del carrito

+++ src/app/carrito/component/carrito.component.ts
@@ -15,16 +15,16 @@ import {
   MatRowDef,
   MatTable
 } from '@angular/material/table';
-import {MatFormField} from '@angular/material/form-field';
-import {MatInput} from '@angular/material/input';
 import {MatChip} from '@angular/material/chips';
-import {Entidad} from '../../permiso/entidad';
+import {Entidad, getEntidadNombre} from '../../permiso/entidad';
 import {MatButton, MatIconButton} from '@angular/material/button';
 import {ItemPedido} from '../../pedido/item-pedido.model';
 import {CurrencyPipe} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {AuthService} from '../../auth/service/auth.service';
 import {Accion} from '../../permiso/accion';
+import {CantidadComponent} from '../../components/cantidad/cantidad.component';
+import {capitalize} from '../../utils/utils';
 
 @Component({
   selector: 'app-carrito',
@@ -35,8 +35,6 @@ import {Accion} from '../../permiso/accion';
     MatHeaderCell,
     MatCell,
     MatColumnDef,
-    MatFormField,
-    MatInput,
     MatTable,
     MatChip,
     MatButton,
@@ -46,7 +44,8 @@ import {Accion} from '../../permiso/accion';
     CurrencyPipe,
     MatHeaderRowDef,
     MatRowDef,
-    FormsModule
+    FormsModule,
+    CantidadComponent
   ],
   templateUrl: './carrito.component.html',
   standalone: true,
@@ -70,24 +69,28 @@ export class CarritoComponent implements OnInit {
     this.cargarCarrito();
   }
 
-  async onCantidadChange(item: ItemPedido) {
-    const cantidad = item.cantidad;
-    if (cantidad < 1) {
-      this.notificationService.show('La cantidad debe ser al menos 1.');
-      return;
-    }
+  async onCantidadChange(item: ItemPedido, cantidad: number) {
+    item.loading = true;
     try {
       await this.pedidoService.updateItemQuantity(item.id, cantidad);
+      item.loading = false;
     } catch {
-      this.notificationService.show('Error al actualizar la cantidad.');
+      this.notificationService.show('Error al actualizar la cantidad');
+      item.loading = false;
     }
   }
 
   async eliminarItem(item: ItemPedido) {
+    if (item.loading) {
+      return;
+    }
+    item.loading = true;
     try {
       await this.pedidoService.removeItem(item.id);
     } catch {
       this.notificationService.show('Error al eliminar el item.');
+    } finally {
+      item.loading = false;
     }
   }
 
@@ -101,4 +104,6 @@ export class CarritoComponent implements OnInit {
 
   protected readonly Entidad = Entidad;
   protected readonly Accion = Accion;
+  protected readonly getEntidadNombre = getEntidadNombre;
+  protected readonly capitalize = capitalize;
 }
